requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
coverage==5.0.1
Django==3.0.1
djangorestframework==3.11.0
mock==3.0.5
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.8
pytest==5.3.2
pytest-cov==2.8.1